+++ client/src/features/common/ObjectSetHierarchy.tsx
@@ -1,8 +1,10 @@
 import React from 'react'
 
 import IEntity from '../../types/data/IEntity'
-import SetParser from '../../lib/parse/data/SetParser'
-import { getNextSetUri } from '../../lib/util/hierarchyHelpers'
+import {
+  getNextSetUris,
+  isEntityAnArchive,
+} from '../../lib/util/hierarchyHelpers'
 
 import GenericBreadcrumbHierarchy from './GenericBreadcrumbHierarchy'
 
@@ -11,17 +13,12 @@ interface IProps {
   columnClassName?: string
 }
 
-function isArchive(entity: IEntity): boolean {
-  const setParser = new SetParser(entity)
-  return setParser.isArchive()
-}
-
 const ObjectSetHierarchy: React.FC<IProps> = ({ entity, columnClassName }) => (
   <GenericBreadcrumbHierarchy
     entity={entity}
     columnClassName={columnClassName}
-    getNextEntityUri={getNextSetUri}
-    linkFilter={isArchive}
+    getNextEntityUri={getNextSetUris}
+    linkFilter={isEntityAnArchive}
     maxLength={8}
   />
 )